requirements.txt
Flask==1.0.2
Routes==2.4.1
psycopg2-binary==2.7.7
nltk==3.4
numpy==1.16.1
nose==1.3.7
pandas==0.24.1
tornado==5.1.1
wordcloud==1.5.0
scikit-learn==0.20.2
rq-dashboard==0.3.12
rq==0.13.0
Flask-Script==2.0.6
